# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: EMMA API
  version: 0.1.0
  description: OpenAPI specification for the EMMA API (developer-docs snapshot)
servers:
  - url: http://localhost:5262
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InteractionCreateDto:
      type: object
      properties:
        type:
          type: string
        direction:
          type: string
          enum: [inbound, outbound]
        subject:
          type: string
        content:
          type: string
        consentGranted:
          type: boolean
        occurredAt:
          type: string
          format: date-time
    InteractionReadDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        type: { type: string }
        subject: { type: string, nullable: true }
        content: { type: string, nullable: true }
        analysisSummary: { type: string, nullable: true }
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
paths:
...
  /api/contacts/{contactId}/interactions:
    get:
      summary: List interactions for a contact
      tags: [Interaction]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: contactId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InteractionReadDto'
              example:
                - id: "44444444-4444-4444-4444-444444444444"
                  timestamp: "2025-08-21T17:00:00Z"
                  type: "call"
                  subject: "Intro call"
                  content: "Discussed onboarding"
        '404': { description: Not Found }
        '400': { description: Bad Request, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '401': { description: Unauthorized, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    post:
      summary: Log an interaction for a contact
      tags: [Interaction]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: contactId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionCreateDto'
            example:
              type: "call"
              direction: "inbound"
              subject: "Intro call"
              content: "Talked through product fit"
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  interactionId: { type: string, format: uuid }
                  analysisPresent: { type: boolean }
                  traceId: { type: string }
              example:
                interactionId: "55555555-5555-5555-5555-555555555555"
                analysisPresent: false
                traceId: "0HMO1B6G4D1A1:00000002"
        '400': { description: Bad Request, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '401': { description: Unauthorized, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
        '500': { description: Error, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
...